RF: Move rules for specification
